feat: gotten a working tool

+++ main.py
@@ -1,11 +1,10 @@
 #!/usr/bin/env python
 
-import requests
 import os
+from tqdm import tqdm
 import click
 import shutil
 from src.utils import get_image_url, download_page, get_number_of_chapters
-import sys
 
 BASE_URL = "https://www.mangasee123.com/manga/"
 BASE_IMG_URL = "https://hot.leanbox.us/manga/"
@@ -13,24 +12,33 @@ BASE_IMG_URL = "https://hot.leanbox.us/manga/"
 @click.command()
 @click.option('--name', default="vinland saga", help = "Name of manga you want to download (default: Vinland saga)")
 @click.option('--start', default = 1, help = "First chapter to download")
-@click.option('--last', default = 10, help = "Last chapter to download")
-@click.option('--save_dir', default = os.path.join(os.getcwd(), "manga"), help="Path to save the manga")
+@click.option('--end', default = 10, help = "Last chapter to download")
+@click.option('--save_dir', default = os.path.join(os.getcwd(), "manga"),
+              help="Absolute path to save the manga. If not specified, a dir `manga` will be created in the pwd and \
+                  the manga will be saved in a sub folder with the name of the manga e.g manga/vinland-saga")
 
-def main(name, start, last, save_dir) -> None:
-    no_chapters = get_number_of_chapters(BASE_URL, name)
-    sys.exit(0)
+def main(name, start, end, save_dir) -> None:
+    no_chapters, formatted_name = get_number_of_chapters(BASE_URL, name)
+
+    save_dir = os.path.join(save_dir, formatted_name)
     
+    if end > no_chapters:
+        print(f"{formatted_name} only has {no_chapters} chapters, {end} exceeds that")
+        raise ValueError("Last chapter exceeds the number of chapters in the manga")
     
     if os.path.exists(save_dir):
         print(f"Path: {save_dir} already exists, deleting it...")
         shutil.rmtree(save_dir)
+        
     os.makedirs(save_dir)
     print("Created: {}".format(save_dir))
+    print("-" *50, end="\n\n")
         
-    for chapter_num in range(start, last + 1):
+    print(f"Dowloading chapters {start} to {end} of {formatted_name}".title())
+    
+    for chapter_num in tqdm(range(start, end + 1), desc="Chapters Downloaded", ncols=100):
         chapter_path = os.path.join(save_dir, f"chap-{chapter_num}")
         os.mkdir(chapter_path)
-        print(f"Created the chapter {chapter_num} directory. Downloading pages...")
         page_num = 1
         
         while True:
@@ -42,6 +50,7 @@ def main(name, start, last, save_dir) -> None:
                 page_num += 1
             else:
                 break
+    print("\nDownload Completed!")
